Testing CameraController script Methods
