+++ Assets/Prefabs/Workspace/Multimeter.prefab
@@ -41,6 +41,7 @@ GameObject:
   m_Component:
   - component: {fileID: 7147250962146585287}
   - component: {fileID: 7147250962146585286}
+  - component: {fileID: 5062015933004513741}
   m_Layer: 0
   m_Name: Multimeter
   m_TagString: Untagged
@@ -74,6 +75,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: de8d1b213c2bbe34aae251af2735b1c6, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &5062015933004513741
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7147250962146585288}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 32994f8bd53a8a543a60f97c2d79e272, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &8258814980005619575
